Incorporate discussions from Trac 166
Thanks for the feedback - I've tried to incorporate it into this newer commit.
Concerning the individual concerns:

1. This was already done in the previous commit, may have been obfuscated by
the fact that I also edited the README to point to the new proposal. This has
made a conflict which I'll resolve in the next commit.
2. I remember seeing this in the agenda for last year's workshop but somehow
the ticket had slipped my mind. I've incorporated the content (in my opinion)
into the proposal, the only things that I've removed where retained for
backwards compatibility. This probably won't be possible for CF-2 so I didn't
attempt it, but if you don't agree I'm happy to talk about it.
3. I've also explicitly referenced the ticket in the proposal so that we have a
trace back to the original document. It may still be worth pursuing using
unsigned types in CF1 separately, but but it made more sense to me to highlight
this as a separate one.